Revisited links in navigation
The previous iteration of links was linking the each nav row as a block. This broke down when we had nested links as well as didn't seem to fit when we added notes in the medium screen.

We have revisited the issue and have gotten closer to an ideal solution. In this PR:

- Added mixins for nav-info icons and content
- Moved to common mixin for all of nav-info--application since redundancy was getting tedious outside of the mixin
- Moved from overarching `a` to multiple `a's` at each level of nav-info

There have been no visual changes