Added HTTP server (#48)
We are formalizing our prototype to use non-HTML-esque URLs (e.g. `/schedule`). The first step to doing that is running an HTTP server. In this PR:

- Added `express` HTTP server
- Added tests
- Documented tests

**Notes:**

Not included in this PR but leading up to it, we did research on Node.js libraries to verify that our long term goals (e.g. job queues, SQL integration) would be plausible and well-tested.