Transferred prototype pages to HTTP server (#49)
To make our prototype more accurate, we want to navigate the UI as though it were live. This PR moves our prototype pages to their intended endpoints (e.g. `/schedule`, `/login`). In this PR:

- Transferred prototype pages to HTTP server (e.g. transferred mock data, moved everything to blocks)
- Added/documented `nodemon` integration
- Repaired image hosting at `dist` for normal and vedor images
- Updated visual tests

**Notes:**

We are going to be adding smoke tests (i.e. verify we can hit the endpoint) to our test suite (visual tests aren't run during CI). We skipped them in this PR to keep the size down