Repaired moment.tz invocations which were using system timezone (#57)
We have been using our system timezone for some of our `moment` instances. This wasn't easily noticable since our OS was on `America/Chicago`. However, now that we moved to Vagrant and `Etc/UTC`, it's much more obvious.

In this PR:

- Repiared bad `moment.tz` invocations

**Notes:**

This impacts screenshots on `/application/:id` and `/interview/:id`