Added new application name behavior (#119)
We have decided to move to a more Google Docs like title approach for application names. This means the name is at the top of the page in an editable input field. The field coming after "Posting URL" felt strange (which would be necessary due to auto-fill).

For naming new applications, we decided to try out auto-fill only for naming. We'll see if it works well or not in production. In this PR:

- Moved from name as field in form to field in subheading
- Added autosize plugin
- Added visual tests

**Before:**

![selection_063](https://cloud.githubusercontent.com/assets/902488/19746809/eb11fc82-9b8c-11e6-8de0-5a7265b71291.png)

**After:**

![selection_064](https://cloud.githubusercontent.com/assets/902488/19746806/e870d796-9b8c-11e6-99d7-832f7113be87.png)