Moved to company names for application names (#127)
We had an issue with using "Role at Company" for application names on small screens; they would get cut off very quickly. To remedy that, we have decided to design around using company names as the application (we are assuming title isn't too valuable to people but we'll see). In this PR:

- Moved to company names for application names
- Updated spacing in UI
- Upgraded to patched `autosize-input` which support `border-box`

**Before:**

![selection_071](https://cloud.githubusercontent.com/assets/902488/20026749/9d746054-a2bd-11e6-8fe1-101ec00eb4da.png)
![selection_072](https://cloud.githubusercontent.com/assets/902488/20026748/9a768896-a2bd-11e6-8603-02052d815d9e.png)

**After:**

![selection_073](https://cloud.githubusercontent.com/assets/902488/20026745/9798762a-a2bd-11e6-9877-8d986b1afb40.png)
![selection_074](https://cloud.githubusercontent.com/assets/902488/20026743/94bbdabe-a2bd-11e6-99a3-44e18701414b.png)