Added timezone locale updater (#145)
We actually don't the leafs of `tz-locales` in our UI but as part of future proofing, we are adding an auto-updater as it (mostly because we wrote it and didn't realized we didn't have to). In this PR:

- Added timezone locale updater

**Notes:**

As mentioned in the initial description, this isn't used in the UI so there are no screenshot changes.