Moved to scenario.route for controller tests (#149)
As mentioned in #147, we have a large API footprint which makes it easy to miss ACL tests (which makes it easy to miss ACL enforcement). To guarantee we are ACL mindful, we added enforcement via `scenario.route`. In this PR:

- Moved all controller tests except `error-handlers` to new ACL paradigm
    - We skipped `error-handlers` as it doesn't fit too well since it's not really a controller

**Notes:**

This PR is quite small when we ignore whitespace changes. To see those, look at:

```
git diff -w
```

or

https://github.com/twolfson/find-work-app/pull/149/files?w=1