Added 404 and login requirement support (#150)
As of #149, we have ACL enforcement in our tests. Our prototype is set up to handle login requirements and 404's already so it makes sense to enable those tests. In this PR:

- Added 404 and login requirement to appropriate controllers
- Enabled/Updated server tests

**Notes:**

There is a decent amount of indentation changes in this PR. Ignore them via:

```
git diff -w
```

or

https://github.com/twolfson/find-work-app/pull/150/files?w=1