Copied navigation content to schedule/archive pages (#159)
We are preparing to revamp our navigation. However, we currently jammed together the navigation and a page's content (schedule/archive). To make changes not be super noisy, we are copying out the content first. In this PR:

- Copied navigation content to schedule/archive pages

![selection_097](https://cloud.githubusercontent.com/assets/902488/20953980/18174dfc-bbec-11e6-9650-abf2198255d2.png)

**Notes:**

The schedule page's content is a bit quirky due to having nav bindings (e.g. it will do hover states) but this isn't meant to be deployed yet