Added slide out menu (#160)
We are transitioning to our new layout via multiple PRs. This PR takes care of adding slide out behavior. The next PR (or 2) should hopefully update everything to new styles. In this PR:

- Added collapse menu functionality
    - We used a custom implementation as it was the fastest to do. Hopefully we didn't ruin the a11y experience (it looks good to me)
- Updated server tests
- Skipped visual tests

**Notes:**

There's a large jump in content on `layout.jade` due to moving the meat of the navigation to a mixin. This was to make it easier to focus on the menu itself rather than its content.

**Screenshots:**

We aren't done with our transition so we won't provide before/after yet. These are our to record our progress so far:

![selection_098](https://cloud.githubusercontent.com/assets/902488/20996363/fbea956e-bcb1-11e6-910c-bf0ef572a557.png)
![selection_099](https://cloud.githubusercontent.com/assets/902488/20996364/fd3cc91e-bcb1-11e6-896b-23fecf9ecc62.png)
![selection_100](https://cloud.githubusercontent.com/assets/902488/20996365/fe557206-bcb1-11e6-81f3-d348da85fe67.png)
![selection_101](https://cloud.githubusercontent.com/assets/902488/20996368/ffeac1a2-bcb1-11e6-9557-101bd516fbc6.png)
![selection_102](https://cloud.githubusercontent.com/assets/902488/20996369/011e442c-bcb2-11e6-8f29-acc7fc0d8c3f.png)