Updated navigation menu (#161)
We have transitioned to our new navigation menu. We aren't quite done with the transition as our schedule/archive content is out of whack as well as some pages should be left-aligned now but we are almost there! In this PR;

- Added `recentlyViewedApplications` mock data
- Fixed support using cookies on mobile
- Updated navigation menu
- Cleaned up tests

**Screenshots:**

Once again, there is no before/after as we are still in transition:

![selection_105](https://cloud.githubusercontent.com/assets/902488/21035878/0baefd76-bd77-11e6-8463-172e1ededac6.png)
![selection_106](https://cloud.githubusercontent.com/assets/902488/21035879/0ca98750-bd77-11e6-84e9-06c20d18bb66.png)
![selection_107](https://cloud.githubusercontent.com/assets/902488/21035881/0e092a24-bd77-11e6-8b16-9694902e3a42.png)