Added common model loader (#168)
We have been planning adding a common model loader for a while. Now we actually need it for nav loading as we need to know about the current model so we can add it to the recent applications array. In this PR:

- Added common model loader and added it to all controllers
- Added server tests

**Notes:**

The API is a bit bulky/clunky but it does the trick and has some nice enforcement (e.g. `layout.jade` will throw if there was no common model loader usage -- ideally that would be in a proxies `get`/`etc` route binding but this is good enough and nav specific which is :+1:)

**Update:**

We improve API cleanliness by switching from `.concat()` inside of controllers to using `_.flatten` so there's no interjected `.concat()`; it looks like a normal Express controller chain