Renamed reminders to be consistent with application status (#175)
We are transferring reminders to their own models so we can start converting mocks to database tables. However, we wanted to make our data more consistent/predictable so we decided to rename the internal names of reminders to be 1:1 with their application status. This also future proofs us for easier renames if we ever have multiple reminders for a single status type.

In this PR:

- Renamed reminders to be consistent with application status (e.g. `offer-response` - > `received-offer`)
- Updated `MOMENT_TZ` tests to use non-reminder field as that's going to change shortly again (i.e. moving data into reminder model)