Moved ids from UUID to VARCHAR column (#180)
We have discovered that we can't do queries with non-UUID values on a UUID (e.g. `/application/404` will 500 due to PostgreSQL error instead of 404). After weighing our options, we decided to weaken the database a little for more flexibility and less overhead in the server of adding param checks in every controller (with the possibility of playing whack-a-mole every time the issue pops up).

In this PR:

- Added migration to convert to VARCHAR
- Added custom data type to distinguish `STRING(36)` from ids
- Added `isUUID` validation to every non-foreign key id location
- Added/updated server tests