Moved to child reminder tables (#189)
We were working on data loading which started to run into issues due to poor cascading of foreign keys. Then, we started to look into cascading foreign keys which started to run into issues due to lacking proper foreign key setup for reminders to their "parent" table. As a result of that, we began to re-evaluate and correct our reminder table to properly have "parent" table so the database automatically scrubs them on deletion of a parent.

After consideration of options like:

- Ignore the issue (bad because we have no data now so it's easily the easiest time)
- Add 2 child tables which only share content via server
    - Downside is easy to get them out of sync, need to make 2 queries for loading all reminders (e.g. on cron)
- Add PostgreSQL inheritance
    - Downside is it's yet another vendor buy in (already using JSON columns for audit logs)
    - Not directly supported by Sequelize

After some time to think, we went with PostgreSQL inheritance as it's unlikely we are going to move off of PostgreSQL in the near future and if we do move off, it should be easy to cast data from child tables to their own tables.

In this PR:

- Added migration for removing `parent_id`/`parent_type` and adding `ApplicationReminder`/`InterviewReminder` tables
- Updated controllers/middlewares/mocks to use `ApplicationReminder`/`InterviewReminder` instead of `Reminder`
- Added inheritance structure for `Reminder` and its child tables including corrections to fixes to relationships
- Added/updated model tests

**Notes:**

We verified that this will work with the database by pulling over some proof of concept changes temporarily

![feels-good.png](https://cloud.githubusercontent.com/assets/902488/21597946/e3167d86-d107-11e6-8dc4-1c19a0980ce3.png)