Added boolean and moment-timezone support to multidict (#192)
We are adding in form validation which requires re-rendering the page with submitted content. To make fallback logic as simple as possible, we are directly exposing the multidict to the view (as it could want a string or array).

We have stickier cases like moments too so we have decided to add `moment` resolving to the multidict itself. This will assist in both the controller creation/update logic as well as in the view and keep us consistent.

In this PR:

- Added `get/fetchBoolean` and `get/fetchMomentTimezone` support to multidict
- Added date/time normalization for `moment-timezone`
- Added server tests