Added data loading for `/application/:id`, `/application/:id/add-application`, `/interview/:id`, and nav (#201)
We started adding data loading support for `/application/:id` but quickly got into a broken set of tests due to nav. We decided to cascade the data loading support for `/application/:id/add-application` and `/interview/:id` as it's light, we still have no live data, and fuller tests are coming soon. In this PR:

- Added data loading support for `/application/:id`, `/application/:id/add-application`, `/interview/:id`, and nav
- Added aliases for fixtures as their names so we can get models when we only care about them as a model (not their status)
- Updated tests