Added saveModelsViaCandidate utility (#210)
We proof of concept'd application edit a second time and wrote `res.saveModels`. We are implementing application actions (e.g. status changes) and want to use that logic so we are porting it over now.

We have changed the implementation to a utility though as we realized we might want to re-render with a validation error but outside of saving (e.g. requiring name if URL fetch fails).

In this PR:

- Added `saveModelsViaCandidate` utility
- Updated application add controller to use `saveModelsViaCandidate`
- Updated server tests
- Added in always using `bluebird` instead of ES6 promise so we have methods like `Promise.all` and consistent error throw behavior