Repaired reminder associations (#212)
While building out action saves into controllers on Friday, we found a scary regression where we would get our "saved for later" reminder as also our "waiting for response" and "received offer" reminder. This was being caused by misconfigured Sequelize.

On Monday, we attempted to fix it via patching over `hasOne` (see comments on this PR) but it turns out we had to move to `belongsTo` even though it isn't semantically accurate:

https://github.com/sequelize/sequelize/pull/7115

In this PR:

- Moved from `hasOne` to `belongsTo` for parent to child reminder ids
- Added tests to prevent future regressions

**Note:**

While this was an ACL issue, it was relatively safe as it used `application.id = reminder.application_id` so there was no chance of exposing other user's data -- only the current user's wrong data.