Added interview type column (#227)
We are about to integrate interview update/delete to our database. Before we get too far though, we want to fix a timely issue where interviews can be "upcoming" (i.e. not yet processed by queue) yet in the past (i.e. time has passed them by). To resolve that, we are introducing the `interview_type` column.

https://trello.com/c/PvJFQ3Od/246-update-interviews-to-have-interview-type-which-we-update-on-change-and-it-cascades-to-application

In this PR:

- Added migration for adding/backfilling interview type column
- Added interview type column to model
- Added validation/automatic hooks to update interview type
- Added/updated server tests