Added past interview support for `/add-application/upcoming-interview` (#228)
We had an edge case where `/add-application/upcoming-interview` could receive a past interview during creation. We could have refused to support this but it's fully plausible to be done by the user in another order (i.e. create waiting for response, then add past interview) so it makes sense to support it here.

In this PR:

- Added past interview support for `/add-application/upcoming-interview`
- Added validation to application/interview to verify upcoming interviews and application are always in sync (i.e. we have upcoming status if and only if there's an upcoming interview)
- Added server tests