Added application reminder support (#245)
We have completed application reminder support and prototyped interview reminder support (but left it out of this PR due to its complexity). In this PR:

- Added application reminder support (i.e. saved for later, waiting for response, received offer)
- Added tests

**Notes:**

We have left out multiple batches and interview reminders as they will only make the PR larger. They will be coming in follow up PRs