Moved to multi-level inheritance for fixtures (#247)
For our existing reminder tests, we have fixtures that were related but not bound to each other. This had potential for fixtures to get out of sync (and as a result tests). To avoid that, we decided to add more complexity for fixture inheritance in exchange for multi-level inheritance. In this PR:

- Moved to multi-level inheritance for fixtures

**Notes:**

For easier PR viewing, please use a whitespace ignorant diff:

https://github.com/twolfson/find-work-app/pull/247/files?w=1