Added `saveModelsViaQueue` and `saveModelsViaServer` helpers (#249)
We are adding support to update interview/application status inside of our queue. We are about to use another `transaction`/`save` combo but realized it was overkill for our needs. Instead, we cleaned up some tech debt and added the helper methods we wanted a while ago.

In this PR:

- Added `saveModelsViaQueue` and `saveModelsViaServer` helpers
- Removed non-utility based `transactions` in server