Moved to normalized routes in analytics (#254)
We have been recording analytics with full UUID strings in them. This works partially but isn't helpful for overall insights. In this PR:

- Moved to normalized routes in analytics

**Notes:**

We could have used Express' internal routes but we have a fear that a simple internal-only change could bump analytics (e.g. renaming `:id` to `:uuid`) as well as using regexp. We went with an external normalizer for sanity