Added time-aware reminder undos (#256)
During the development of db-attached reminders, we noticed that the existing undo behavior wasn't intuitive. This PR is our attempt to correct that. The logic stands that:

- If I recently moved off of application status, then I should reuse that original reminder since it was likely accidental
    - Must be within 5 minutes
    - Reminder must not be past-due
- If I didn't recently move off of application status (i.e. we might have accidentally received offer a long time ago), then I should generate a new reminder

Our logic achieves this by touching active reminders on status changes and then checking the to-be restored reminder's `updated_at` timestamp. In this PR:

- Added active reminder touching before changes occur
- Added `updated_at`/expiration timestamp checking for reminders
- Added/updated tests