Updated logo and favicon (#260)
We received some feedback a while ago about our favicon looking like a steering wheel. We had similar personal feelings. We explored alternative logos again in December but wanted to let our thoughts grow on them as well as focus on more important tasks. Now that we are approaching launch, it seems appropriate to perform the logo update.

In this PR:

- Updated logo and favicon
- Updated visual test signatures

**Before:**

![selection_213](https://cloud.githubusercontent.com/assets/902488/22846853/909e4b50-ef9e-11e6-8cac-40cbdeee6d90.png)
![selection_214](https://cloud.githubusercontent.com/assets/902488/22846855/91fbfb64-ef9e-11e6-914e-cfcb76070ef8.png)

**After:**

![selection_211](https://cloud.githubusercontent.com/assets/902488/22846858/96f87e80-ef9e-11e6-87ac-97a2593f3c8e.png)
![selection_212](https://cloud.githubusercontent.com/assets/902488/22846861/98b3621c-ef9e-11e6-9c6e-72b4c6d43fea.png)