Added browser tests (#269)
We are about to add AJAX loading to our research company pages. Since AJAX can get finnicky, we want to be able to test it. As a result, we are adding in Karma and backfilling pending tests. In this PR:

- Added Karma
- Added browser tests
- Added PhantomJS dependency to Wercker Dockerfile
- Upgraded Node.js while performing Wercker update

**Notes:**

We waffled on our syntax for a bit -- trying to get something as consistent as possible between `server`/`browser` without losing out on things like XSS. I think we found a nice compromise. For reference, here's our first shot:

```
describe('An pair of content-synced elements with content', function () {
  it('initializes with target element\'s content', function () {
    // Create our content
    var containerEl = h.div([
      h.input({
        attrs: {'data-content-sync': '#target'},
        value: 'foo'
      }),
      h.p('#target')
    ]);

    // Initialize our content and assert
    contentSync.init(containerEl);
    expect($(containerEl).find('#target').text()).to.equal('foo');
  });
});
```