Removed TODOs (#288)
We have accrued a lot of TODOs over the lifetime of our application. TODOs are annoying as they aren't too searchable/prioritizable since they aren't in Trello (with most of our tasks). This PR removes a lot of TODOs by either deleting them or moving them to Trello tasks. The TODOs that remain are all linked to Trello tasks so they should be searchable/prioritizable.

In this PR:

- Removed TODOs
- Removed legacy `grid` integration -- we didn't use this after our first few days iirc =/