Replaced AngelList results with external links (#309)
We have come to the realization that an AngelList integration will take 3 or 4 days (2 days for email, 2 days for AngelList OAuth) and we definitely want to launch soon. We were unable to find any reasonable APIs to replace it (LinkedIn is our org only, Crunchbase has little useful info). As a result, we settled on linking to searches instead. In this PR:

- Replaced AngelList results with external links
- Updated tests
- Updated landing page to reflect external links

**Before:**

![selection_277](https://cloud.githubusercontent.com/assets/902488/23642621/dd6ce21e-02b0-11e7-8509-f2777b3cdd9b.png)
![selection_278](https://cloud.githubusercontent.com/assets/902488/23642622/de5f70d8-02b0-11e7-89be-8a22ad00b3a4.png)

**After:**

![selection_279](https://cloud.githubusercontent.com/assets/902488/23642617/d9659bd4-02b0-11e7-8312-6d550f4d43ba.png)
![selection_281](https://cloud.githubusercontent.com/assets/902488/23642614/d5ba2be4-02b0-11e7-8be2-efa8db12a767.png)