Added indirect event analytics (#312)
We have started to get traffic and users signing up on our site. We want to better track the source of where good users come from. To do this, we are adding indirect event analytics so we can track sign up/log in. In this PR:

- Added indirect event analytics
- Added tests

![selection_290](https://cloud.githubusercontent.com/assets/902488/23817017/f8c6614c-05a4-11e7-80e5-ce151cf9d58b.png)
![selection_289](https://cloud.githubusercontent.com/assets/902488/23817009/ebbdfcc6-05a4-11e7-8ce2-66131b6162ec.png)

**Notes:**

We could probably track some more events (e.g. application create/save/delete) but we are going to keep it to sign up for now so this gets into the wild ASAP