Moved to common auth view/controller (#313)
We are adding email login and want to cut down on maintenance work. To reduce such redundancy, we are consolidating our view/controller for login/sign up into a single file with some if/else logic. We feel that the content is consistent enough now to warrant it.

We could move to a single route as well with "authenticate" wording but feel it's best to keep the dual functionality illusion for now in case we want to undo it later.

In this PR:

- Moved to common view for sign up/login
- Moved to common controller for sign up/login