Add limit to GC size (#326)
We have had a recurring issue where production exceeds its memory usage due to what looked like a memory leak. After being unable to reproduce with forced `gc()` calls, we realized it's most likely GC not wanting to run and not a memory leak. As a result, we are adding a limit to GC size to prevent these false positives.

In this PR:

- Added limit to GC size in `queue`/`server`
- Added note in `package.json`

**Notes:**

This feels like reasonable information to share so we'll likely write up a blog post